+++ Harvest.Web/ClientApp/src/Util/StringFormatting.ts
